Split the user.js file into their own commands
