+++ Panel/bot/discord/commands/user/help.js
@@ -1,0 +1,9 @@
+exports.run = async (client, message, args) => {
+    const embed = new Discord.MessageEmbed()
+        .addField("__**Commands**__`", config.DiscordBot.Prefix + "user new` | Create an account \n`" +
+            config.DiscordBot.Prefix + "user password` | Reset account password \n`" +
+            config.DiscordBot.Prefix + "user link` | Link this account with console account \n`" +
+            config.DiscordBot.Prefix + "user unlink` | Unlinks account from the console account \n`" +
+            config.DiscordBot.Prefix + "user premium` | Check your premium server limit")
+    await message.channel.send(embed)
+}
