+++ Panel/bot/discord/commands/user/unlink.js
@@ -1,0 +1,4 @@
+exports.run = async (client, message, args) => {
+    userData.delete(message.author.id)
+    message.channel.send('You have unlinked this account!')
+}
